+++ headers/sequencer.h
@@ -11,49 +11,151 @@
 #include <list>
 #include <memory>
 
-//once this returns true, we stop running it
-struct RunThis
+class Sequencer
 {
+public:
     typedef std::function<bool(int)> Func;
-    Func func; //function to run, takes in an int that is the number of times this has been run
+private:
+    struct SequenceNode;
+    typedef std::shared_ptr<SequenceNode> RunPtr;
+    typedef std::weak_ptr<SequenceNode> WeakRunPtr;
 
-    int ranNumTimes = 0; //number of times this function has been called
 
-    RunThis(const Func& f) : func(f)
+    //A linked list-esque wrapper around a callable function
+    struct SequenceNode
     {
 
+        RunPtr next;
+        Func func; //function to run, takes in an int that is the number of times this has been run
+
+        int ranNumTimes = 0; //number of times this function has been called
+
+        SequenceNode(const Func& f) : func(f)
+        {
+            
+        }
+        virtual bool operator()()
+        {
+            bool val = func(ranNumTimes);
+            ranNumTimes++;
+            return val;
+        }
+    };
+
+    RunPtr head;
+    WeakRunPtr end; 
+public:
+    Sequencer(Func func){add(func);};
+    Sequencer(){};
+
+    RunPtr getHead(){return head;}
+    WeakRunPtr getEnd(){return end;}
+    
+    Sequencer& add(Sequencer other)
+    {
+        if (!other.empty())
+        {
+            if (!empty())
+            {
+                end.lock()->next = other.getHead();
+            }
+            else
+            {
+                head = other.getHead();
+            }
+            end = other.getEnd();
+        }
+        return *this;
+    }
+    Sequencer& add(Func func)
+    {
+        SequenceNode* node = new SequenceNode(func);
+        if (!empty())
+        {
+            end.lock()->next.reset(node);
+            end = end.lock()->next;
+        }
+        else
+        {
+            head.reset(node);
+            end = head;
+        }
+        return *this;
+    }
+    Sequencer& parallel(Func newFunc)
+    {
+        if (!empty())
+        {
+            //modify our end-most node's function to call this new function as well
+            end.lock()->func = [newFunc,oldFunc=end.lock()->func](int times)
+                                {
+                                    bool val1 = oldFunc(times);
+                                    bool val2 = newFunc(times); //gotta call them separately so we don't short circuit
+                                    return val1 && val2;
+                                }; 
+        }
+        else
+        {
+            add(newFunc);
+        }
+
+        return *this;
     }
 
-    bool operator()()
+    void pop()
     {
-        bool val = func(ranNumTimes);
-        ranNumTimes++;
-        return val;
+        if (head)
+        {
+            //move to the next node
+            //possibly delete the current node if there are no other pointers pointing to it
+            head = head->next;
+        }
+    }
+
+    bool empty() const
+    {
+        return (head.get() == nullptr);
+    }
+
+    //clear the sequence by setting the head to past the last node
+    void clear() 
+    {
+        if (!empty())
+        {
+            head = end.lock()->next;
+            end = head;
+        }
+    }
+
+    bool run()
+    {
+        if (empty())
+        {
+            return true;
+        }
+        else
+        {
+            return (*head)();
+        }
     }
 };
 
-//runs the front most function until it returns true, at which point is is removed
-typedef std::list<RunThis> Sequencer;
 
-//an actual sequence
-//runs until empty. If a Sequence is the only pointer to a Sequencer, it is deleted
-//however, Sequencers with multiple pointers are kept, even if empty. This allows "queueing", where an object (like GameCamera),
-//can continuously add RunThis to one Sequence
-typedef std::shared_ptr<Sequencer> SequencePtr;
+//runs the front most function until it returns true, at which point is is removed
+//typedef std::list<RunThis> Sequencer;
+
 //stores Sequencers, which can be set to run either once per frame or once per loop
 struct Sequences
 {
-    typedef std::list<SequencePtr> SequencerList;
+    typedef std::list<std::shared_ptr<Sequencer>> SequencerList;
     static SequencerList physicsSequences; //sequences that run once per frame
     static SequencerList renderSequences; //sequences that run once per loop
 
     static void run(SequencerList& lst);
-    static bool run(SequencePtr& seq); //run a sequence, return true if it is done (has no more units to run)
 
 public:
+
     //isPhysics = true means we add to "physicsSequences", else "renderSequences"
-    static SequencePtr add(Sequencer* sequence, bool isPhysics);
-    static SequencePtr add(const RunThis& runThis, bool isPhysics);
     /**
      * @brief Waits until the provided function returns true
      * 
@@ -61,19 +163,40 @@ public:
      * @param isPhysics true if physics
      * @return SequencePtr 
      */
-    static SequencePtr waitFor(std::function<bool()> runThis,bool isPhysics);
-    static void add(SequencePtr& seq, bool isPhysics);  //add a Sequence, that already exists, probably because you want to have multiple pointers
-
-    template<typename... Callable>
-    static SequencePtr add(bool isPhysics, Callable... callable)
-    {
-        return add((new Sequencer{RunThis::Func(callable)...}),isPhysics);
-    }
+    static std::shared_ptr<Sequencer>&  add(bool isPhysics, Sequencer::Func func);
+    static std::shared_ptr<Sequencer>&  add(bool isPhysics, Sequencer&& seq);
+    static std::shared_ptr<Sequencer>&  add(bool isPhysics, std::shared_ptr<Sequencer>& seq);
 
     static void runPhysics();
     static void runRenders();
 };
 
+template<auto... Initializers>
+struct InitFunc
+{
+    std::tuple<std::invoke_result_t<decltype(Initializers)>...> values;
+    //the actual function we will call
+    //there is totally a way to do this without having to make it a member and instead having it be an auto template parameter
+    //However, the parameter pack nature of Initializers would mean this template parameter would have to be the first template parameter
+    //For readability, I really like having the function at the end since it uses the Initializer's return values
+    std::function<bool(std::invoke_result_t<decltype(Initializers)>...,int)> func;
+
+    InitFunc(decltype(func) func_) : func(func_){}
+
+    bool operator()(int times)
+    {
+        if(times == 0)
+        {
+            init();
+        }
+        return std::apply(func,std::tuple_cat(values,std::make_tuple(times))); 
+    }
+private:
+    void init()
+    {
+        values = std::make_tuple(Initializers()...);
+    }
+};
 
 
 #endif // SEQUENCER_H_INCLUDED