+++ src/terrain.cpp
@@ -50,7 +50,7 @@ void ObjectLookup::clear()
 
 bool GlobalTerrain::isValidObject(PhysicsBody* obj, LayerType layer)
 {
-    return obj && obj->orient.layer == layer && !obj->isDead();
+    return obj && obj->getOrient().layer == layer && !obj->isDead();
 }
 
 void GlobalTerrain::addObject(std::shared_ptr<PhysicsBody> ptr, LayerType layer)
@@ -218,18 +218,32 @@ void GlobalTerrain::render()
     //render layers from back all the way to front, not including layers past the camera
     for (int i = layers.size() - 1; i >= limit ;i--)
     {
-        layers[i].terrain.render(i - Globals::Game.getCurrentLayer(),getZOfLayer(i));
-
+        float z = Globals::Game.terrain.getZOfLayer(i);
+        if (i == limit)
+        {
+            //if there are any objects that are past the current layer (at this point, just the player when portalling),
+            //render them before objects
+            for (auto it = layers[limit].objects.begin(); it != layers[limit].objects.end(); ++it)
+            {
+                if (PhysicsBody* obj = it->lock().get())
+                if (obj->getOrient().getZ() > z)
+                {
+                    obj->render();
+                }
+            }
+        }
+        layers[i].terrain.render(i,z);
     }
-
-    //only render entities of the current layer
+    //render the rest of the objects
     for (auto it = layers[limit].objects.begin(); it != layers[limit].objects.end(); ++it)
     {
-        if (PhysicsBody* obj = it->lock().get()) [[likely]] //highly likely since update removes bad pointers and it always runs first
+        if (PhysicsBody* obj = it->lock().get()) //highly likely since update removes bad pointers and it always runs first
         {
             obj->render();
         }
     }
+
+
 
 }